requirements.txt
openai>=1.12.0
pydantic>=2.6.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
tiktoken>=0.5.0
tenacity>=8.2.0
sentence-transformers>=2.2.0
rich>=13.0.0
prompt-toolkit>=3.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-cov>=4.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0
sqlalchemy>=2.0.0
psycopg2-binary
pgvector
python-jose[cryptography]
python-multipart
pytest-playwright

# Dev tools
pre-commit>=4.0.0
ruff>=0.9.0
pyright>=1.1.390
